Fixed -v to actually show colors
The previous version was using the wrong escape code to show colors. I've implemented a version that simply draws a line in each color by setting only the background color. This could be made more complex by also setting the foreground color. (I've included a setfg function, just in case, but the tests I tried were not aesthetically pleasing). 

Fixing the color showing functionality also required adding in some extra doohickies to prevent possible bugs. For example, if you've got 256 colors, it should be piped through a pager so it doesn't scroll by too fast. Also, if the user hits ^C in the middle of printing in color, the terminal should still reset to the default colors. And finally, this patch adds the ability to request a certain number of colors to show instead of using the counted number. 

I have some doubts about this patch as it increases the file size greatly and seems to conflate showing colors with counting colors. It might make more sense to have xterm-color-count.sh only count colors and then have a new script, xterm-color-show.sh which calls the counting script and shows all the possible colors.